backend/minio_watcher/requirements.txt
minio==7.2.8
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
